// src/types/nansen/api.gen.ts
export const S_EntitySearchResult = z.looseObject({
  "name": z.string(),
  "tags": z.array(z.string()).nullish(),
  "rank": z.number().nullish(),
});
// ...
export const S_TokenSearchResult = z.looseObject({
  "name": z.string(),
  "symbol": z.string(),
  "chain": z.string(),
  "address": z.string(),
  "price": z.number().nullish(),
  "volume_24h": z.number().nullish(),
  "market_cap": z.number().nullish(),
  "rank": z.number().nullish(),
});
export type TokenSearchResult = z.infer<typeof S_TokenSearchResult>;
export const S_GeneralSearchResponse = z.looseObject({
  "tokens": z.array(S_TokenSearchResult).nullish(),
  "entities": z.array(S_EntitySearchResult).nullish(),
  "total_results": z.number(),
});
export type GeneralSearchResponse = z.infer<typeof S_GeneralSearchResponse>;